+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # floo-rtc-ucloud-android Changelog
 [蓝莺IM](https://www.lanyingim.com/)，是由美信拓扑团队研发的新一代即时通讯云服务，SDK设计简单集成方便，服务采用云原生技术和多云架构，私有云也可按月付费。
 
+## 1.0.2 - 2023/08/06
+- 优化：修改UCloud环境配置到properties文件。
+- 修改README.md，增加UCloud环境配置说明。
+
 ## 1.0.1 - 2023/08/04
 新功能：RTC一对一视频及语音通话。